# Speed up the computation process