## Tasks

* [x] Make use of `from_le` and `to_le`
* [x] Get rid of auxiliary buffers
* [x] Do not initialize `buffer`
* [x] Make use of `copy_nonoverlapping`
* [ ] ~Make use of `cfg!(target_endian = "…")`~

## Before

```
rustc 1.35.0-nightly (befeeb7c0 2019-03-30)
rustup 1.17.0 (069c88ed6 2019-03-05)
cargo 1.35.0-nightly (63231f438 2019-03-27)

test compute_0001000 ... bench:       3,006 ns/iter (+/- 72)
test compute_0010000 ... bench:      29,242 ns/iter (+/- 594)
test compute_0100000 ... bench:     289,585 ns/iter (+/- 9,280)
test compute_1000000 ... bench:   2,922,448 ns/iter (+/- 88,887)
```

## After

```
rustc 1.35.0-nightly (befeeb7c0 2019-03-30)
rustup 1.17.0 (069c88ed6 2019-03-05)
cargo 1.35.0-nightly (63231f438 2019-03-27)

test compute_0001000 ... bench:       2,420 ns/iter (+/- 49)
test compute_0010000 ... bench:      23,645 ns/iter (+/- 290)
test compute_0100000 ... bench:     235,915 ns/iter (+/- 5,451)
test compute_1000000 ... bench:   2,362,815 ns/iter (+/- 41,799)
```