@IvanUkhov 2019-03-31
@jens1o, thanks for your interest! I was hoping to get the code review before merging. However, now I’m thinking if this is something worth doing at all. The speed-up is not that impressive. It’s around 20%, as one can see above. I am not sure if this justifies this much of unsafe code.