@IvanUkhov 2020-01-31
@hellow554, thank you for taking the time to review it! It was probably a bad idea right from the start. The observed speed-up is not worth the effort. My motivation was that such command-line tools as `md5sum` are blazingly fast compared to this package, which is a bit annoying 🙂